delete experience is bulletproofed

+++ controllers/experienceController.js
@@ -151,21 +151,28 @@ const deleteExperience = async (req, res) => {
                     message: "Experience does not exist!",
                 });
             } else {
-                Experience.deleteOne(result)
-                    .then((deleteResult) => {
-                        console.log(deleteResult);
-                        res.send({
-                            success: true,
-                            message: "Experience deleted",
-                        });
-                    })
-                    .catch((err) => {
-                        console.log(err);
-                        res.send({
-                            success: false,
-                            message: err,
-                        });
+                if (result.userId != req.body.userId) {
+                    res.send({
+                        success: false,
+                        message: "User does not have that skill",
                     });
+                } else {
+                    Experience.deleteOne(result)
+                        .then((deleteResult) => {
+                            console.log(deleteResult);
+                            res.send({
+                                success: true,
+                                message: "Experience deleted",
+                            });
+                        })
+                        .catch((err) => {
+                            console.log(err);
+                            res.send({
+                                success: false,
+                                message: err,
+                            });
+                        });
+                }
             }
         });
     }
